requirements-dev.txt
# Testing Dependencies
pytest>=8.3.0
pytest-cov>=6.0.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
pytest-timeout>=2.3.0
requests-mock>=1.12.0

# Linting and Code Quality
black>=24.10.0
flake8>=7.1.0
pylint>=3.3.0
mypy>=1.13.0
bandit>=1.8.0

# Type Stubs
types-requests>=2.31.0
types-python-dateutil>=2.8.19

# Code Coverage
coverage>=7.3.0
codecov>=2.1.13

# Testing Utilities
faker>=19.3.1
freezegun>=1.2.2
responses>=0.23.3

